requirements.txt
semantic-version==2.3.0
Fabric==1.8.2
PyYAML==3.10
boto==2.28.0
python-dateutil==2.2
unittest2==0.5.1
glob2==0.4.1
jenkinsapi==0.2.20